Debug site: fix SEO, dark mode, responsive, and HTML issues
- nav.html: remove duplicate id="themeToggle" (HTML spec violation)
- sections.css: dark mode hero dot-grid, landscape phone hero compact layout, 320px breakpoint
- journey.html + careers.html: add JSON-LD structured data (BreadcrumbList, WebPage, Organization)
- pages/privacy.html + terms.html: add viewport-fit=cover and meta description
- robots.txt: add Bingbot, Slurp, DuckDuckBot, Baiduspider, YandexBot
- sitemap.xml: update lastmod dates to today
- js/services.js: parseFloat instead of parseInt for CSS var unit stripping
- js/form.js + css/components.css: replace inline opacity with .btn-loading class

+++ js/services.js
@@ -14,7 +14,7 @@
 
   const subnavEl = document.querySelector('.svc-subnav');
   const subnavH = subnavEl ? subnavEl.offsetHeight : 48;
-  const navH = parseInt(getComputedStyle(document.documentElement).getPropertyValue('--nav-height')) || 68;
+  const navH = parseFloat(getComputedStyle(document.documentElement).getPropertyValue('--nav-height')) || 68;
 
   const spy = new IntersectionObserver(entries => {
     entries.forEach(entry => {
